examples/langchain/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
chromadb>=0.4.22
python-dotenv>=1.0.0
openai>=1.3.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.5.0
termcolor>=2.3.0
